Look for "required" required under "schema".
We had been looking for "required" at the LDO level, and
copying it into the parameters object (which would blank
out any "required" present inside the top level of "schema").

This removes that copy and adds a regression test for the
old behavior.  It also renames a test variable to be more
clear (a linkDefinition would be expected to be an LDO,
which made that test appear to pass by looking at an LDO
when the fixture actually had "required" inside of "schema",
which was confusing when debugging the issue).