Remove no-longer-needed "ID" hack.
At some point, some library somewhere would incorrectly treat
properties called "id" as JSON Schema "id" keywords, which was
never correct behavior.  We compensated for it by recognizing
"ID" and downcasing it.  This is no longer needed.

The comment mentioned a "preserveCase" option but I cannot find
any other mention of it in the code or tests, so perhaps that
was never implemented.