Replace nonstandard rel: self with cfRecurse
It was getting horribly confusing to keep having to distinguish
between '"rel":"self"-in-the-LDO' and
'our-nonstandard-"rel":"self"-in-link-schemas', so make the
nonstandard one into its own keyword instead of overloading "rel".

"cfRecurse" takes a plain JSON Pointer (not a URI) which is
evaluated from the root of the post-$ref-resolution JSON object.

Note that the root JSON Pointer is an empty string!!!

Currently only the root JSON Pointer is supported (which covers
the non-standard "rel": "self" case).  Adding support for
more is fairly trivial if needed but let's wait for a use case.

Also fix a unit test that was not actually being run due to
indentation confusion.