Support link mediaType and encType
APIs can send other content types such as XML, JavaScript files
and many other things as requests and responses.

Currently, we only support application/json (the default) and
multipart/form-data (which has a special syntax for curl).

This adds support for any other media type, with the expectation
that, when "encType" or "mediaType" is used this way,
"schema" or "targetSchema" (respectively) will only have
an "example" field with a pre-formatted string.

We detect these situations and avoid passing the example through
the JSON formatter.  This required a recognizing example-only
schemas directly, as they would otherwise produce an empty
object example (the default behavior of the old code).